Add conditional checks to graphpass & cat jobs; resolves #104
- Add conditional to cat job
- Add conditional to graphpass job
- Update rubocop config

+++ app/jobs/collections_cat_job.rb
@@ -13,11 +13,21 @@ class CollectionsCatJob < ApplicationJob
       combine_full_url_output_cmd = 'cat ' + collection_derivatives + '/all-domains/output/part* > ' + collection_derivatives + '/all-domains/' + c.collection_id.to_s + '-fullurls.txt'
       logger.info 'Executing: ' + combine_full_url_output_cmd
       system(combine_full_url_output_cmd)
-      FileUtils.rm_rf(collection_derivatives + '/all-domains/output')
+      successful_job = collection_derivatives + '/all-domains/' + c.collection_id.to_s + '-fullurls.txt'
+      if File.exist? successful_job
+        FileUtils.rm_rf(collection_derivatives + '/all-domains/output')
+      else
+        raise 'Collections cat domains job failed.'
+      end
       combine_full_text_output_cmd = 'cat ' + collection_derivatives + '/all-text/output/part* > ' + collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
       logger.info 'Executing: ' + combine_full_text_output_cmd
       system(combine_full_text_output_cmd)
-      FileUtils.rm_rf(collection_derivatives + '/all-text/output')
+      successful_job = collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
+      if File.exist? successful_job
+        FileUtils.rm_rf(collection_derivatives + '/all-text/output')
+      else
+        raise 'Collections cat full text job failed.'
+      end
     end
   end
 end
